Moved instanceOf and updated call sites
