+++ base/shared/src/main/scala/scalaz/core/package.scala
@@ -1,7 +1,5 @@
 package scalaz
 
 package object core {
-
   val Void: VoidModule = VoidImpl
-
 }
