+++ base/shared/src/main/scala/scalaz/ct/kleisli.scala
@@ -3,6 +3,7 @@ package ct
 
 import scala.inline
 
+import kernel.instanceOf
 import algebra.MonoidClass
 import data.~>
