+++ base/shared/src/main/scala/scalaz/ct/strong.scala
@@ -3,6 +3,8 @@ package ct
 
 import scala.language.experimental.macros
 
+import kernel.instanceOf
+
 trait StrongClass[P[_, _]] extends ProfunctorClass[P] {
   def first[A, B, C](pab: P[A, B]): P[(A, C), (B, C)]
   def second[A, B, C](pab: P[A, B]): P[(C, A), (C, B)]
