+++ base/shared/src/main/scala/scalaz/data/disjunction.scala
@@ -3,6 +3,7 @@ package data
 
 import scala.{ inline, Either }
 
+import kernel.instanceOf
 import core.EqClass
 import ct._
 import debug.DebugClass
