+++ base/shared/src/main/scala/scalaz/data/identity.scala
@@ -1,6 +1,7 @@
 package scalaz
 package data
 
+import kernel.instanceOf
 import core.EqClass
 import ct.MonadClass
 import debug.DebugClass
