+++ base/shared/src/main/scala/scalaz/data/ilist.scala
@@ -1,6 +1,7 @@
 package scalaz
 package data
 
+import kernel.instanceOf
 import core.EqClass
 import debug.DebugClass
 import types.IsCovariantClass
