+++ base/shared/src/main/scala/scalaz/data/upstar.scala
@@ -3,7 +3,8 @@ package data
 
 import scala.AnyVal
 
-import scalaz.ct.ProfunctorClass
+import kernel.instanceOf
+import ct.ProfunctorClass
 
 final case class UpStar[F[_], A, B](run: A => F[B]) extends AnyVal
