+++ base/shared/src/main/scala/scalaz/external/io.scala
@@ -1,8 +1,9 @@
 package scalaz
-package core
+package external
 
 import java.lang.Throwable
 
+import kernel.instanceOf
 import algebra.MonoidClass
 import ct._
 import types.IsCovariantClass
@@ -62,3 +63,4 @@ trait FiberInstances {
       }
     })
 }
+
