+++ base/shared/src/main/scala/scalaz/external/package.scala
@@ -1,0 +1,5 @@
+package scalaz
+
+package object external {
+  import acyclic.pkg
+}
