+++ base/shared/src/main/scala/scalaz/std/utils.scala
@@ -1,6 +1,7 @@
 package scalaz
 package std
 
+import kernel.instanceOf
 import core.EqClass
 import data.Cord
 import debug.DebugClass
