+++ base/shared/src/main/scala/scalaz/types/iscovariant.scala
@@ -1,6 +1,8 @@
 package scalaz
 package types
 
+import kernel.instanceOf
+
 /**
  * Witnesses that the type constructor `F[_]` is covariant,
  * even though the variance annotation of its type parameter has been forgotten.
