+++ project/ScalazBuild.scala
@@ -45,8 +45,11 @@ object Scalaz {
       libraryDependencies ++= compileOnlyDeps ++ testDeps ++ Seq(
         compilerPlugin("org.spire-math"         %% "kind-projector"  % "0.9.7"),
         compilerPlugin("com.github.tomasmikula" %% "pascal"          % "0.2.1"),
-        compilerPlugin("com.github.ghik"        %% "silencer-plugin" % "1.0")
+        compilerPlugin("com.github.ghik"        %% "silencer-plugin" % "1.0"),
+        compilerPlugin("com.lihaoyi"            %% "acyclic"         % "0.1.7"),
+        "com.lihaoyi" %% "acyclic" % "0.1.7" % "provided"
       ),
+      autoCompilerPlugins := true,
       incOptions ~= (_.withLogRecompileOnMacro(false))
     ) ++ {
       Seq(packageBin, packageDoc, packageSrc).flatMap {
