+++ std/shared/src/main/scala/collection/list.scala
@@ -5,6 +5,7 @@ import scala.{ ::, List, Nil }
 
 import scala.Predef.$conforms
 
+import scalaz.kernel.instanceOf
 import core.EqClass
 import ct.MonadClass
 import data.Cord
